requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
